requirements.txt
Django~=4.1.10
python-dotenv~=1.0.0
Scrapy~=2.9.0
pymongo~=3.12.3
praw~=7.7.1
pandas~=2.0.3
dnspython==2.4.2
djongo==1.3.6
sqlparse==0.2.4
openai==0.27.9
minio==7.1.16
pyspark==3.4.1
elasticsearch==8.9.0
confluent_kafka==2.2.0
django-cors-headers==4.2.0
requests==2.31.0
cassandra-driver==3.28.0